Enum mariadb::storage::StorageError

pub enum StorageError {

    KeyNotFound = 120,
    FoundDuppKey = 121,
    InternalError = 122,
    RecordChanged = 123,
    WrongIndex = 124,
    Crashed = 126,
    WrongInRecord = 127,
    OutOfMem = 128,
    RetryInit = 129,
    NotAtAble = 130,
    WrongCommand = 131,
    OldFile = 132,
    NoActiveRecord = 133,
    RecordDeleted = 134,
    RecordFileFull = 135,
    IndexFileFull = 136,
    EndOfFile = 137,
    Unsupported = 138,
    ToBigRow = 139,
    GCreateOption = 140,
    FoundDuppUnique = 141,
    UnknownCharset = 142,
    WrongMrgTableDef = 143,
    CrashedOnRepair = 144,
    CrashedOnUsage = 145,
    LockWaitTimeout = 146,
    LockTableFull = 147,
    ReadOnlyTransaction = 148,
    LockDeadlock = 149,
    CannotAddForeign = 150,
    NoReferencedRow = 151,
    RowIsReferenced = 152,
    NoSavepoint = 153,
    NonUniqueBlockSize = 154,
    NoSuchTable = 155,
    TableExist = 156,
    NoConnection = 157,
    NullInSpatial = 158,
    TableDefChanged = 159,
    NoPartitionFound = 160,
    RbrLoggingFailed = 161,
    DropIndexFk = 162,
    ForeignDuplicateKey = 163,
    TableNeedsUpgrade = 164,
    TableReadonly = 165,
    AutoincReadFailed = 166,
    AutoincErange = 167,
    Generic = 168,
    RecordIsTheSame = 169,
    LoggingImpossible = 170,
    CorruptEvent = 171,
    NewFile = 172,
    RowsEventApply = 173,
    Initialization = 174,
    FileTooShort = 175,
    WrongCrc = 176,
    TooManyConcurrentTrxs = 177,
    NotInLockPartitions = 178,
    IndexColTooLong = 179,
    IndexCorrupt = 180,
    UndoRecTooBig = 181,
    InvalidDocid = 182,
    TablespaceExists = 184,
    TooManyFields = 185,
    RowInWrongPartition = 186,
    RowNotVisible = 187,
    AbortedByUser = 188,
    DiskFull = 189,
    IncompatibleDefinition = 190,
    FtsTooManyWordsInPhrase = 191,
    DecryptionFailed = 192,
    FkDepthExceeded = 193,
    TablespaceMissing = 194,
    SequenceInvalidData = 195,
    SequenceRunOut = 196,
    CommitError = 197,
    PartitionList = 198,
    NoEncryption = 199,

}

Storage handler error types, as defined in my_base.h

Variants

KeyNotFound = 120

Didn’t find key on read or update

FoundDuppKey = 121

Duplicate key on write

InternalError = 122

Internal error

RecordChanged = 123

Update with is recoverable

WrongIndex = 124

Wrong index given to function

Crashed = 126

Indexfile is crashed

WrongInRecord = 127

Record-file is crashed or table is corrupt

OutOfMem = 128

Out of memory

RetryInit = 129

Initialization failed and should be retried

NotAtAble = 130

not a MYI file - no signature

WrongCommand = 131

Command not supported

OldFile = 132

old databasfile

NoActiveRecord = 133

No record read in update()

RecordDeleted = 134

A record is not there

RecordFileFull = 135

No more room in file

IndexFileFull = 136

No more room in file

EndOfFile = 137

end in next/prev/first/last

Unsupported = 138

unsupported extension used

ToBigRow = 139

Too big row

GCreateOption = 140

Wrong create option

FoundDuppUnique = 141

Duplicate unique on write

UnknownCharset = 142

Can’t open charset

WrongMrgTableDef = 143

conflicting tables in MERGE

CrashedOnRepair = 144

Last (automatic?) repair failed

CrashedOnUsage = 145

Table must be repaired

LockWaitTimeout = 146

LockTableFull = 147

ReadOnlyTransaction = 148

Updates not allowed

LockDeadlock = 149

CannotAddForeign = 150

Cannot add a foreign key constr.

NoReferencedRow = 151

Cannot add a child row

RowIsReferenced = 152

Cannot delete a parent row

NoSavepoint = 153

No savepoint with that name

NonUniqueBlockSize = 154

Non unique key block size

NoSuchTable = 155

The table does not exist in engine

TableExist = 156

The table existed in storage engine

NoConnection = 157

Could not connect to storage engine

NullInSpatial = 158

NULLs are not supported in spatial index

TableDefChanged = 159

The table changed in storage engine

NoPartitionFound = 160

There’s no partition in table for given value

RbrLoggingFailed = 161

Row-based binlogging of row failed

DropIndexFk = 162

Index needed in foreign key constr

ForeignDuplicateKey = 163

Upholding foreign key constraints would lead to a duplicate key error in e other table.

TableNeedsUpgrade = 164

The table changed in storage engine

TableReadonly = 165

The table is not writable

AutoincReadFailed = 166

Failed to get next autoinc value

AutoincErange = 167

Failed to set row autoinc value

Generic = 168

Generic error

RecordIsTheSame = 169

row not actually updated: new values same as the old values

LoggingImpossible = 170

It is not possible to log this statement

CorruptEvent = 171

The event was corrupt, leading to illegal data being read

NewFile = 172

New file format

RowsEventApply = 173

The event could not be processed. No other handler error happene.

Initialization = 174

Error during initialization

FileTooShort = 175

File too short

WrongCrc = 176

Wrong CRC on page

TooManyConcurrentTrxs = 177

oo many active concurrent transactions

NotInLockPartitions = 178

There’s no explicitly listed partition in table for the given value

IndexColTooLong = 179

Index column length exceeds limit

IndexCorrupt = 180

Index corrupted

UndoRecTooBig = 181

Undo log record too big

InvalidDocid = 182

Invalid InnoDB Doc ID

TablespaceExists = 184

The tablespace existed in storage engine

TooManyFields = 185

Table has too many columns

RowInWrongPartition = 186

Row in wrong partition

RowNotVisible = 187

AbortedByUser = 188

DiskFull = 189

IncompatibleDefinition = 190

FtsTooManyWordsInPhrase = 191

Too many words in a phrase

DecryptionFailed = 192

Table encrypted but decrypt failed

FkDepthExceeded = 193

FK cascade depth exceeded

TablespaceMissing = 194

Missing Tablespace

SequenceInvalidData = 195

SequenceRunOut = 196

CommitError = 197

PartitionList = 198

NoEncryption = 199